backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
aiomysql==0.2.0
sqlalchemy==2.0.23
alembic==1.13.1

# 文件处理
aiofiles==23.2.1
python-magic==0.4.27

# AI和向量处理
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3

# 缓存
redis==5.0.1
aioredis==2.0.1

# 工具库
httpx==0.25.2
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
werkzeug==2.3.7